+++ CHANGES.rst
@@ -12,11 +12,14 @@
 - Various improvements to the online documentation. [#1400, #1425]
 - Fixed a bug in ``tpf.interact()`` so that proper y label is displayed when
   lightcurve is normalized with ``transform_func``. [#1387]
+- Changed 'short' cadence in search.py to be <200s so the new TESS FFI cadence is excluded [#1394]
 - Fixed ``lightkurve.utils.centroid_quadratic()`` in edge cases, e.g., fluxes are
   all negative with mask specified, NaN in the identified brightest 3X3 patch. [#1426]
 - Fixed interact features, e.g. ``tpf.interact()``, to be compliant
   with Bokeh v3.4.0. The minimum Bokeh version is raised to v2.3.2 accordingly. [#1428]
 - Updated aperture plotting to speed it up in the case of large (bleed column) apertures. [#1434]
+- Fixed time format for KEPSEISMIC light curves from bjkd to mjd. [#1443]
+- Fixed flux unit bug in to_fits [#1454]
 
 
 2.4.2 (2023-11-03)
@@ -33,6 +36,7 @@
 - Made `LombScarglePeriodogram` compatible with Astropy v5.3 [#1342]
 - Updated the TPF plotting function to work correctly with WCS plotting [#1298]
 - Added the ability to open light curves from the TGLC High Level Science Product
+- Added in a flux_err mask to remove NaNs in pldcorrector
 
 2.4.0 (2023-02-14)
 ==================